imgur_client/requirements.txt
imgurpython
Pillow
celery